+++ autotools/cli.py
@@ -42,6 +42,7 @@ def cli(ctx, perf):
             - autounit: Convert measurement units (meters to feet, liters to gallons, etc)
             - autozip: Compress files and directories into various archive formats (zip, tar.gz, etc)
             - autotodo: Create and manages a simple task list in a markdown file
+            - autonote: Takes quick notes and saves them to a markdown file
             - test: Run the test suite (development only)
 
         \b